combined block and preview block, need to fix curr # error and then propagation
