+++ src/components/Block/Block.scss
@@ -1,17 +1,6 @@
 $valid_bg: #b4ff9a;
 $invalid_bg: #ff9a9a;
 
-.block-background {
-  border-radius: 10px;
-  padding: 1rem 2rem;
-}
-
-.bordered-background {
-  border: 1px dashed black;
-  border-radius: 10px;
-  padding: 0 0.5em 0.5em;
-}
-
 .block {
   width: 20rem;
 }
